+++ bin/align_rnaseq_gsnap.pl
@@ -38,6 +38,7 @@ Optional:
  -memory             Memory for samtools sorting, use suffix G M b (def '35G')
  -verbose
  -piccard_0m         Ask gsnap to add 0M between insertions (only for piccard compatibility, issues with most other software)
+ -filetype       :s  Only process files ending with this text
 
 =head1 AUTHORS
 
@@ -83,7 +84,7 @@ my $intron_length      = 70000;
 my $cpus               = 6;
 my $memory             = '35G';
 my $pattern1            = '_1_';
-
+my $filetype = '';
 &GetOptions(
              'debug'           => \$debug,'verbose'=>\$verbose,
              'fasta:s'         => \$genome,
@@ -104,7 +105,8 @@ my $pattern1            = '_1_';
              'commands_only:s' => \$just_write_out_commands,
              'split_input:i'   => \$split_input,
              'notpaired'       => \$notpaired,
-             'piccard_0m'      => \$piccard_0m
+             'piccard_0m'      => \$piccard_0m,
+	     'filetype:s'      => \$filetype,
 );
 
 pod2usage if $help;
@@ -135,7 +137,7 @@ unless ( $pattern2 || $notpaired ) {
  $pattern2 =~ s/1/2/;
 }
 
-my @files = glob("$input_dir/*$pattern1*");
+my @files = glob("$input_dir/*$pattern1*$filetype");
 push( @files, "$input_dir/$pattern1" ) if -s "$input_dir/$pattern1";
 push( @files, @ARGV );
 my %verified_files;
@@ -325,8 +327,7 @@ sub checked_paired_files() {
    chomp($lines);
    my $number_of_lines = int( ( $lines / 4 ) / $split_input );
    $number_of_lines *= 4;
-   die
-"Number of lines is not as expected for FASTQ ($number_of_lines / $lines)\n"
+   die "Number of lines is not as expected for FASTQ ($number_of_lines / $lines)\n"
      unless $number_of_lines % 4 == 0;
    system("split -a 3 -d -l $number_of_lines $file $file. ");
    system("split -a 3 -d -l $number_of_lines $file $pair. ");
@@ -384,9 +385,8 @@ sub align_unpaired_files() {
   &process_cmd( $file_align_cmd, '.', "gsnap.$base*" )
     unless (    -s "$base_out_filename"."_uniq"
              || -s "$base_out_filename"."_uniq.bam" );
-  unless ( -s "$base_out_filename"."_uniq.bam" ) {
-   &process_cmd(
-"$samtools_exec view -h -u -T $genome $base_out_filename"."_uniq | $samtools_exec sort -@ $samtools_sort_CPUs -l 9 -m $memory -o $base_out_filename"."_uniq.bam -");
+  unless ( -s "$base_out_filename"."_uniq.bam" || $just_write_out_commands) {
+   &process_cmd("$samtools_exec view -h -u -T $genome $base_out_filename"."_uniq | $samtools_exec sort -@ $samtools_sort_CPUs -l 9 -m $memory -o $base_out_filename"."_uniq.bam -");
    &process_cmd("$samtools_exec index $base_out_filename"."_uniq.bam");
 
    ## For JBrowse
@@ -400,7 +400,7 @@ sub align_unpaired_files() {
    );
    unlink("$base_out_filename"."_uniq");
   }
-  unless ( -s "$base_out_filename"."_mult.bam" ) {
+  unless ( -s "$base_out_filename"."_mult.bam" || $just_write_out_commands) {
    &process_cmd("$samtools_exec view -h -u -T $genome $base_out_filename"."_mult | $samtools_exec sort -@ $samtools_sort_CPUs -l 9 -m $memory -o $base_out_filename"."_mult.bam -");
    &process_cmd("$samtools_exec index $base_out_filename"."_mult.bam");
    print LOG "\n$base_out_filename"."_mult.bam:\n";
@@ -417,7 +417,7 @@ sub align_unpaired_files() {
 #   &process_cmd("$samtools_exec flagstat $base_out_filename"."_uniq_mult.bam >> gsnap.$base.log"   );
 #  }
 
-  print LOG "\nGSNAP Completed!\n";
+  print LOG "\nGSNAP Completed!\n" unless $just_write_out_commands;
   close LOG;
   print CMD " rm -f $file " if $just_write_out_commands && $split_input;
   print CMD "\n" if $just_write_out_commands;
@@ -461,10 +461,9 @@ sub align_paired_files() {
   $file_align_cmd .= $qual_prot if $qual_prot;
 
   $file_align_cmd .=    " --split-output=gsnap.$base --read-group-id=$base $file $pair ";
-  &process_cmd( $file_align_cmd, '.', "gsnap.$base*" )
-    unless (    -s "$base_out_filename"."_uniq"
-             || -s "$base_out_filename"."_uniq.bam" );
-  unless ( -s "$base_out_filename"."_uniq.bam" ) {
+  &process_cmd( $file_align_cmd, '.', "gsnap.$base*" )    unless (    -s "$base_out_filename"."_uniq" || -s "$base_out_filename"."_uniq.bam" );
+
+  unless ( -s "$base_out_filename"."_uniq.bam" || $just_write_out_commands) {
    &process_cmd("$samtools_exec view -h -u -T $genome $base_out_filename"."_uniq | $samtools_exec sort -@ $samtools_sort_CPUs -l 9 -m $memory -o $base_out_filename"."_uniq.bam -");
    &process_cmd("$samtools_exec index $base_out_filename"."_uniq.bam");
 
@@ -474,12 +473,10 @@ sub align_paired_files() {
    &process_cmd("bedGraphToBigWig $base_out_filename"."_uniq.coverage.bg $genome.fai $base_out_filename"."_uniq.coverage.bw") if `which bedGraphToBigWig`;
 
    print LOG "\n$base_out_filename"."_uniq.bam:\n";
-   &process_cmd(
-    "$samtools_exec flagstat $base_out_filename"."_uniq.bam >> gsnap.$base.log"
-   );
+   &process_cmd( "$samtools_exec flagstat $base_out_filename"."_uniq.bam >> gsnap.$base.log" );
    unlink("$base_out_filename"."_uniq");
   }
-  unless ( -s "$base_out_filename"."_mult.bam" ) {
+  unless ( -s "$base_out_filename"."_mult.bam" || $just_write_out_commands) {
    &process_cmd("$samtools_exec view -h -u -T $genome $base_out_filename"."_mult | $samtools_exec sort -@ $samtools_sort_CPUs -l 9 -m $memory -o $base_out_filename"."_mult.bam -");
    &process_cmd("$samtools_exec index $base_out_filename"."_mult.bam");
    print LOG "\n$base_out_filename"."_mult.bam:\n";
@@ -488,7 +485,7 @@ sub align_paired_files() {
    );
    unlink("$base_out_filename"."_mult");
   }
-  if ( -s $out_halfmapped && !-s "$out_halfmapped.bam"){
+  if (!$just_write_out_commands && (-s $out_halfmapped && !-s "$out_halfmapped.bam")){
     &process_cmd("$samtools_exec view -h -u -T $genome $out_halfmapped | $samtools_exec sort -@ $samtools_sort_CPUs -l 9 -m $memory -o $out_halfmapped.bam -");
     &process_cmd("$samtools_exec index $out_halfmapped.bam");
     unlink($out_halfmapped) if -s "$out_halfmapped.bam";
@@ -500,7 +497,7 @@ sub align_paired_files() {
 #   &process_cmd("$samtools_exec flagstat $base_out_filename"."_uniq_mult.bam >> gsnap.$base.log"   );
 #  }
 
-  print LOG "\nGSNAP Completed!\n";
+  print LOG "\nGSNAP Completed!\n" unless $just_write_out_commands;
   close LOG;
   print CMD " rm -f $file $pair " if $just_write_out_commands && $split_input;
   print CMD "\n" if $just_write_out_commands;
